Avoid using legacy stuff for ebos linear solvers.
