+++ CMakeLists_files.cmake
@@ -279,6 +279,7 @@ list (APPEND PUBLIC_HEADER_FILES
   opm/autodiff/DebugTimeReport.hpp
   opm/autodiff/DuneMatrix.hpp
   opm/autodiff/ExtractParallelGridInformationToISTL.hpp
+  opm/autodiff/FlowLinearSolverParameters.hpp
   opm/autodiff/FlowMain.hpp
   opm/autodiff/FlowMainEbos.hpp
   opm/autodiff/FlowMainSequential.hpp
