+++ opm/autodiff/CPRPreconditioner.hpp
@@ -30,6 +30,7 @@
 #include <opm/common/utility/platform_dependent/disable_warnings.h>
 #include <opm/common/utility/parameters/ParameterGroup.hpp>
 #include <opm/autodiff/ParallelOverlappingILU0.hpp>
+#include <opm/autodiff/FlowLinearSolverParameters.hpp>
 
 #include <dune/istl/bvector.hh>
 #include <dune/istl/bcrsmatrix.hh>
@@ -56,62 +57,6 @@ namespace Opm
 template<typename O, typename S, typename C,
          typename P, std::size_t COMPONENT_INDEX>
 class BlackoilAmg;
-
-/**
- * \brief Parameters used to configure the CPRPreconditioner.
- */
-struct CPRParameter
-{
-    double cpr_relax_;
-    double cpr_solver_tol_;
-    int cpr_ilu_n_;
-    MILU_VARIANT cpr_ilu_milu_;
-    bool cpr_ilu_redblack_;
-    bool cpr_ilu_reorder_sphere_;
-    int cpr_max_ell_iter_;
-    bool cpr_use_amg_;
-    bool cpr_use_bicgstab_;
-    bool cpr_solver_verbose_;
-    bool cpr_pressure_aggregation_;
-
-    CPRParameter() { reset(); }
-
-    CPRParameter( const ParameterGroup& param)
-    {
-        // reset values to default
-        reset();
-
-        cpr_relax_                = param.getDefault("cpr_relax", cpr_relax_);
-        cpr_solver_tol_           = param.getDefault("cpr_solver_tol", cpr_solver_tol_);
-        cpr_ilu_n_                = param.getDefault("cpr_ilu_n", cpr_ilu_n_);
-        cpr_ilu_redblack_         = param.getDefault("ilu_redblack", cpr_ilu_redblack_);
-        cpr_ilu_reorder_sphere_   = param.getDefault("ilu_reorder_sphere", cpr_ilu_reorder_sphere_);
-        cpr_max_ell_iter_         = param.getDefault("cpr_max_elliptic_iter",cpr_max_ell_iter_);
-        cpr_use_amg_              = param.getDefault("cpr_use_amg", cpr_use_amg_);
-        cpr_use_bicgstab_         = param.getDefault("cpr_use_bicgstab", cpr_use_bicgstab_);
-        cpr_solver_verbose_       = param.getDefault("cpr_solver_verbose", cpr_solver_verbose_);
-        cpr_pressure_aggregation_ = param.getDefault("cpr_pressure_aggregation", cpr_pressure_aggregation_);
-
-        std::string milu("ILU");
-        cpr_ilu_milu_ = convertString2Milu(param.getDefault("ilu_milu", milu));
-    }
-
-    void reset()
-    {
-        cpr_relax_                = 1.0;
-        cpr_solver_tol_           = 1e-2;
-        cpr_ilu_n_                = 0;
-        cpr_ilu_milu_             = MILU_VARIANT::ILU;
-        cpr_ilu_redblack_         = false;
-        cpr_ilu_reorder_sphere_   = true;
-        cpr_max_ell_iter_         = 25;
-        cpr_use_amg_              = true;
-        cpr_use_bicgstab_         = true;
-        cpr_solver_verbose_       = false;
-        cpr_pressure_aggregation_ = false;
-    }
-};
-
 
 namespace ISTLUtility
 {
